pyproject.toml
[project]
name = "rag-with-localdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.2.12",
    "langchain>=0.3.11",
    "langchain-community>=0.3.11",
    "langchainhub>=0.1.21",
    "faiss-cpu>=1.9.0.post1",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
]
